requirements.txt
boto3
botocore
psycopg2-binary
SQLAlchemy==1.3.24
SQLAlchemy-Paginator